+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+* Add -httpprof flag to enable HTTP profiling with pprof []()
+
 ### Deprecated
 
 ### Known Issues